Address @nicoinch feedback
